+++ WireDomain/Sources/WireDomain/Repositories/Conversations/Repository/ConversationRepositoryError.swift
