+++ WireDomain/Tests/WireDomainTests/LocalStores/ConversationLocalStoreTests.swift
@@ -23,6 +23,7 @@ import WireDataModelSupport
 import WireDomainSupport
 import WireTestingPackage
 import XCTest
+
 @testable import WireAPI
 @testable import WireDomain
